fix CC issue, split exporter
